Setup no longer requires hardcoded paths to BNG, WESTPA, but does still require that the bngl is in the same directory as the yaml
